# src/main/java/com/ezycollect/paymentapplication/model/ApiResponse.java
package com.ezycollect.paymentapplication.model;

import java.time.LocalDateTime;

public class ApiResponse {
    private int status;
    private String message;

    public ApiResponse(int status, String message) {
        this.status = status;
        this.message = message;
    }

    public int getStatus() {
        return status;
    }

    public String getMessage() {
        return message;
    }

    public void setStatus(int status) {
        this.status = status;
    }

    public void setMessage(String message) {
        this.message = message;
    }
}